move AWS::Elasticsearch::Domain
